feat: add _publication api for the historical versions
