+++ src/server/api/mod.rs
@@ -1,6 +1,8 @@
 //! This module contains the API endpoints for the server.
 pub mod archive;
 pub mod date;
+pub mod doc_transform;
+pub mod notifications;
 pub mod routes;
 pub mod serve;
 pub mod state;
